Rocky 5 algorithm now works, 'but place in ring' improvement doesn't seem to be an improvement for 100+ numbers

+++ libs/libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: dbegara- <dbegara-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/14 18:11:51 by dbegara-          #+#    #+#             */
-/*   Updated: 2021/05/11 17:35:10 by dbegara-         ###   ########.fr       */
+/*   Updated: 2021/05/12 15:44:18 by dbegara-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(char *str)
+size_t	ft_strlen(char *str)
 {
 	int	counter;
